Improve access control for node creation
To check the access policy when creating a new node, the content of the incoming node resource must be used: The logged-in user must be OWNER of the organization where the node is to be added to. In the first implementation, the owner information was accessed on the raw incoming request. This is insecure, the request must be validated first. The approach here lets the existing serializer validate the request before performing the access control check.