First attempt to include the time series of a node in the node detail resource
Working on the node- and installation timeseries got me thinking if it would be better to make the timeseries part of the node resp. installation resource and thus get rid of the standalone timeseries resources.
This commit is a first step in that direction. It provides a node list and node details resouce, where the node list includes the current sample count and the node detail includes the entire time series.