Include the time series of a node installation into the installation resource
This is the continuation of my effort to simplify the time series API. Similar to the node resource, the installation resource now also includes its associated time series if requested via a query parameter. The installation time series are those samples that are recorded by the installed node during the time span of the installation.

Includes tests and an updated API documentation.